Rename oncp_rec_size → partial_rec_size
This variable is used to track the size of partially-received tunnel packets
across mainloop invocations.

It was intially used by the oNCP protocol, where tunneled packets can be
split across TLS records.  We're now going to use it to deal with an
extremely similar mis-layering of packetisation for PPP-based protocols
as well.

Renaming it to partial_rec_size to emphasize its cross-protocol nature (cf.
https://gitlab.com/openconnect/openconnect/-/merge_requests/151).

Signed-off-by: Daniel Lenski <dlenski@gmail.com>
Signed-off-by: David Woodhouse <dwmw2@infradead.org>