Ensure that tunnel configuration script always receives both IPv6 address and netmask, or neither
The logic for handling IPv6 address configuration in vpnc-script is complex
(https://gitlab.com/openconnect/vpnc-scripts/blob/800f671b/vpnc-script#L228-245),
because the script cannot rely on *both* 'INTERNAL_IP6_ADDRESS' *and*
'INTERNAL_IP6_NETMASK' being set.  That is because these variables were
historically set based on the behavior of Cisco servers
(https://gitlab.com/openconnect/openconnect/blob/3d845bc9/cstp.c#L560-567),
which are not guaranteed to send both.

This patch ensures that if we send *one* of these variables to the script,
we will always send *both*, continuing the changes begun in
https://gitlab.com/openconnect/openconnect/-/commit/3d845bc9bf62b3e816e0d2fd72970ef33964a191#e1f68b30df0b044d45c9bd36c0856d78f9682f04:

1. If only the IPv6 netmask is set by the server, we remove the /prefixlen
   to make the IPv6 address.
2. If only the IPv6 address is set by the server, we append /128 to make
   the IPv6 netmask.

Once merged, this will allow us to simplify vpnc-script and make it less
dependent on the historical Cisco-specific behavior (albeit in a way that is
not backwards-compatible with older verisons of OpenConnect).

Signed-off-by: Daniel Lenski <dlenski@gmail.com>