Ensure that tunnel configuration script always receives Legacy IP address and netaddr/netmask, or neither
The logic for handling Legacy IP configuration in vpnc-script is very
complex, confusing, and somewhat redundant (see
https://gitlab.com/openconnect/vpnc-scripts/blob/800f671b/vpnc-script):

1. Set a default Legacy IP route on the tunnel interface *unless*
   split-include routes are provided.
2. Set a route to INTERNAL_IP4_NETADDR/INTERNAL_IP4_NETMASK, if provided
   (redundant unless split-include routes provided).
3. Set split-include and split-exclude routes unconditionally, if provided.

This behavior is based on what Cisco servers send; in particular, they
expect us to set a default Legacy IP route if no split-includes are sent.

This patch ensures that if we send a Legacy IP address to the script, we
will *always* also send netaddr/netmask, and vice versa.

Once merged, this will allow us to simplify vpnc-script and make it less
dependent on the historical Cisco-specific behavior (albeit in a way that is
not backwards-compatible with older verisons of OpenConnect).

Signed-off-by: Daniel Lenski <dlenski@gmail.com>