Use `()` and `{}` instead of `list()` and `dict()`
This fixes a DeepSource alert:

Consider using literal syntax to create the data structure

Using the literal syntax can give minor performance bumps compared to
using function calls to create dict, list and tuple.

This is because here, the name dict must be looked up in the global
scope in case it has been rebound. Same goes for the other two types
list() and tuple().

Signed-off-by: Dimitri Papadopoulos <3350651-DimitriPapadopoulos@users.noreply.gitlab.com>