Add --sni option to the CLI, for domain fronting
This option allows the user to override the hostname value presented as the
SNI (Server Name Indication) for new TLS connections.

This may be useful for Domain Fronting, a technique by which some filtered
or censored Internet connections can be bypassed.

Note that sending different values for the SNI and 'Host:' header violates
HTTP standards, is prevented by many cloud hosting providers, AND will cause
the server to send a certificate for the wrong domain name (thus requiring the
use of '--servercert').

Signed-off-by: Daniel Lenski <dlenski@gmail.com>