If --sni is specified, expect peer certificate to match value sent in SNI, rather than hostname
As David points out in
https://gitlab.com/openconnect/openconnect/-/merge_requests/297#note_718276474,
if the client overrides the SNI sent in the TLS handshake, that *already*
implies that the client expects the server to send a TLS certificate
matching a different hostname.

This patch refactors both the GnuTLS and OpenSSL code slightly, in order to
avoid duplication of the code paths used for verifying that the server's
certificate matches a specific hostname or IP address.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>