Sending --long-options to HIP script was a mistake; use environment variables instead
Sending parameters to HIP script as --long-options was a mistake that was
based on overly-close replication of the invocation of the CSD script/binary
(see auth.c). In the case of CSD, some parameters *need* to be sent on the
command line to maintain compatibility with opaque Cisco CSD binaries.

For GlobalProtect/HIP, we have no need to maintain compatibility with any
opaque binaries sent by the server.

For anything that hasn't already shipped in a released version, we should
use environment variables as the standard way to send values to the HIP
script, particularly because it makes it easier for a shell script to parse
them and accept new ones.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>