Avoid printing spurious ENOENT error from EPOLL_CTL_DEL
In openconnect_close_https() we always unmonitor the ssl_fd even if we are
still in the auth phase and hadn't actually monitored it yet. I think I'd
like to keep track of that better, with a flag to explicitly keep track
of whether a given fd is registered. Or maybe abusing a high bit of the
existing vpninfo->XXX_epoll variables that already keep state.

For the imminent release though, let's take the simple approach and just
ignore the harmless error. Not pretty, but safer.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>