Support vhost on more than just x86_64
The reason we only supported x86_64 is because the vhost interface is
kind of awful and requires that we *know* what TASK_SIZE is.

If we specify a range in with the VHOST_SET_MEM_TABLE ioctl that goes
higher than TASK_SIZE, it doesn't fail immediately, which would be
helpful. Instead it succeeds, but then *all* buffer accesses fail,
even the ones at valid addresses, because our mapping *could* have
pointed to an invalid address for which !access_ok().

So... here's a really awful hack to attempt to determine TASK_SIZE
by attempting to mmap() at different addresses. For addresses which
are invalid, it'll return -ENOMEM, but for addresses which are valid
it returns -EINVAL because we deliberately don't specify either
MAP_PRIVATE or MAP_SHARED as we should. This is better than the criu
version which tries to *unmap*, but in doing so much actually unmap
something that was supposed to be there!

Fixes: #383
Signed-off-by: David Woodhouse <dwmw2@infradead.org>