Handle Fortinet split-exclude routes
Per logs shared in https://gitlab.com/openconnect/openconnect/-/issues/394,
these can be specified on Fortinet 7.0+ with a `<split-tunnel-info
negate="1">` tag in the XML configuration, whereas `<split-tunnel-info/>`
(without the `negate="1"` attribute) will continue to indicate split-INClude
routes.

Specification of split-include routes should (continue to) cause us to omit
a default route for the VPN tunnel, but split-excludes alone should not
trigger this; the routing tables shown in the issue demonstrate this.

Assumptions that seem reasonable but are not explicitly shown in the logs:

1. If *both* split-include and split-exclude routes are specified, the VPN
   config would contain two separate `<split-tunnel-info/>` tags, one with
   the `negate="1"` attribute and one without.
2. IPv6 split-excludes would work the same as Legacy IP split-excludes,
   except for being defined under the `<ipv6/>` tag rather than the
   `<ipv4/>` tag.

This also modifies that XML test configuration for the
`fake-fortinet-server.py` to include both Legacy IPv4 and IPv6
split-excludes, based on the two above assumptions about how these should be
structured.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>