Export STRAP private key with AnyConnect cookie
For STRAP we need to reconnect using the same private key as the auth
process did. Thankfully we already have precedent for this; we can put
multiple 'cookies' into the opaque string that is passed from auth to
connection process, and use internal_split_cookies() to parse them.

So encode the privkey into an 'openconnect_strapkey' cookie which we
handle specially on ingestion.

Fix up a few places where vpninfo->cookie was handled directly, to make
it all work.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>